+++ src/Styles/createjobs.css
@@ -1,0 +1,96 @@
+* {
+  box-sizing: border-box;
+  padding: 0;
+  margin: 0;
+}
+
+/* Job Creator Wrapper */
+.job_creator {
+  width: 100vw;
+  height: 100vh;
+  display: flex;
+  align-items: left;
+  justify-content: space-between;
+  overflow: hidden;
+}
+
+/* Job Tabs */
+.job-tab {
+  width: 80%;
+  height: 70%;
+  min-height: 400px;
+  background: #93cee7;
+  margin: 3.5rem auto 1.5rem;
+  padding: 2rem 1rem;
+  color: #424849;
+  border-radius: 2rem;
+  border-top-left-radius: 0;
+  @media (max-width: 768px) {
+    padding: 2rem 0;
+  }
+}
+
+ul.nav-job-tabs {
+  width: 60%;
+  margin: 0 auto 2rem;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin-top: -84px;
+  margin-left: -16.5px;
+  border: 1px solid #93cee7;
+  border-top-right-radius: 2rem;
+  border-top-left-radius: 2rem;
+  @media (max-width: 768px) {
+    width: 90%
+  }
+}
+
+ul.nav-job-tabs li {
+  width: 50%;
+  margin-right: -16.5;
+  padding: 1rem;
+  list-style: none;
+  text-align: center;
+  cursor: pointer;
+  transition: all 0.7s;
+  border-bottom-left-radius: 0;
+  border-top-left-radius: 2rem;
+  border-top-right-radius: 2rem;
+}
+
+ul.nav-job-tabs li:nth-child(2) {
+  border-radius: 0;
+  border-bottom-right-radius: 0;
+  border-top-left-radius: 2rem;
+  border-top-right-radius: 2rem;
+}
+
+ul.nav-job-tabs:hover {
+  background: rgba(233, 184, 236, 0.168);
+}
+
+ul.nav-job-tabs li.active {
+  background: #93cee7;
+}
+
+.outlet {
+  background: 93cee7;
+  margin: 20px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
+
+.job-form h1, .cronjob-form h1 {
+  font-size: 2rem;
+  text-align: center;
+}
+/* 
+form fieldset {
+  display: flex; 
+  flex-direction: column;
+  justify-content: space-around;
+} */
+
+
